requirements.txt
# RAG-Based Quiz Generator — Dependencies
# Install with: pip install -r requirements.txt

# LLM & Embeddings
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8

# PDF processing
pypdf>=4.0.0

# Vector store
faiss-cpu>=1.8.0

# TF-IDF baseline retriever
scikit-learn>=1.4.0
numpy>=1.26.0

# Frontend
gradio>=4.30.0

# Environment management
python-dotenv>=1.0.0
